==> notebooks/1_tratamento_dados.ipynb <==
# Tratamento Inicial - Análise Exploratória

---

### Este notebook faz parte do portfólio acadêmico do curso de Análise de Dados - EBAC.

#### Neste notebook realizamos a leitura e inspeção inicial do dataset `clientes.csv`.
#### As etapas incluem visualização dos dados, verificação de estrutura, tipos de variáveis e identificação de valores nulos.

## 1. Leitura dos Dados

Importamos bibliotecas essenciais e carregamos o dataset.

import pandas as pd
from pathlib import Path

data_path = Path('dataset/clientes.csv')
df = pd.read_csv(data_path)

df.head()

## 2. Visualização Inicial

-  Exibição das primeiras e últimas linhas completas do DataFrame
> - `head()` - mostra as 5 primeiras linhas da tabela.
> - `tail()` - mostra as 5 últimas linhas.
> - `to_string()` - garante que a visualização seja completa no terminal(sem cortes).

print(df.head().to_string())
print(df.tail().to_string())

## 3. Estrutura do DataFrame

- Verificação da quantidade de linhas e colunas
> - `shape` – retorna uma tupla com(n_linhas, n_colunas)

print('Qtd: ', df.shape)

## 4. Tipagem das Variáveis

- Identificação dos tipos de dados por coluna
> - `dtypes` – mostra se cada coluna é `int`, `float`, `object`(texto), `datetime`, etc.

print('Tipagem: \n', df.dtypes)

## 5. Verificação de Valores Nulos

- Identificação e contagem de dados ausentes
> - `isnull()` – retorna um DataFrame booleano indicando células nulas
> - `sum()` – soma os valores nulos por coluna

print('Valores nulos: \n', df.isnull().sum())
